refactor: grouped Persistence file adapter for plugins in separate folder
