+++ src/Adapters/Driven/Persistence.File/Plugin/IPluginDiscriminatorHolder.cs
@@ -1,4 +1,4 @@
-﻿namespace ScriptBee.Persistence.File;
+﻿namespace ScriptBee.Persistence.File.Plugin;
 
 public interface IPluginDiscriminatorHolder
 {
