+++ src/Adapters/Driven/Persistence.File/Plugin/IPluginManifestYamlFileReader.cs
@@ -1,6 +1,6 @@
 ﻿using ScriptBee.Domain.Model.Plugin.Manifest;
 
-namespace ScriptBee.Persistence.File;
+namespace ScriptBee.Persistence.File.Plugin;
 
 public interface IPluginManifestYamlFileReader
 {
