+++ test/Adapters/Driven/Persistence.File.Tests/Plugin/Manifest/PluginManifestYamlParsingTests.cs
@@ -1,4 +1,5 @@
 ﻿using ScriptBee.Domain.Model.Plugin.Manifest;
+using ScriptBee.Persistence.File.Plugin;
 
 namespace ScriptBee.Persistence.File.Tests.Manifest;
