+++ test/Adapters/Driven/Persistence.File.Tests/Plugin/Manifest/TestData/LoaderPluginManifest.yaml
